+++ lib/@page.dart
@@ -1,17 +1,16 @@
 import 'package:flutter/widgets.dart';
 import 'package:learn_flutter/page.dart';
 
-import 'defalut_layout.dart';
+import 'root_tree.dart';
 
 Meta rootPage = Meta(
   title: "home",
   builder: build,
-  layout: <T>(N<T> note) => DefaultLayout<T>(note),
-
+  layout: <T>(N<T> note) => DefaultLayout<T>(current: note, tree: paths.note),
 );
 
 build(Pen pen, BuildContext context) {
-  pen. markdown(r'''
+  pen.markdown(r'''
 ## 欢迎来到flutter世界
