+++ lib/page.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:learn_flutter/navigator_v2.dart';
-import 'package:learn_flutter/defalut_layout.dart';
 
 import 'utils.dart';
 import 'package:markdown/markdown.dart' as md;
@@ -16,9 +15,10 @@ class Meta<T> {
     required this.title,
     required this.builder,
     LayoutBuilder? layout,
-  }) {
-    this.layoutBuilder = layout ?? <T>(N<T> note) => DefaultLayout<T>(note);
-  }
+  }) : layoutBuilder = layout ??
+            <T>(N<T> note) => DefaultLayout<T>(
+                  current: note,
+                );
 
   Pen<T> build(BuildContext context, N<T> p) {
     //Pen一次性用品，用完丢弃，防止Flutter框架多次刷新造成的状态问题
@@ -122,7 +122,11 @@ class N<T> implements Rule<T> {
   /// 树形父子Page的页面骨架有继承性，即自己没有配置骨架，就用父Page的骨架
   LayoutBuilder get layout {
     if (isRoot) {
-      return meta == null ? <T>(N<T> note) => DefaultLayout<T>(note) : meta!.layoutBuilder;
+      return meta == null
+          ? <T>(N<T> note) => DefaultLayout<T>(
+                current: note,
+              )
+          : meta!.layoutBuilder;
     }
     return meta?.layoutBuilder != null ? meta!.layoutBuilder : _parent!.layout;
   }
@@ -132,6 +136,8 @@ class N<T> implements Rule<T> {
   int get level => isRoot ? 0 : _parent!.level + 1;
 
   bool get isRoot => _parent == null;
+
+  N get root => isRoot ? this : _parent!.root;
 
   @override
   String get path {
@@ -182,3 +188,118 @@ class N<T> implements Rule<T> {
 /// Dart 3 class-modifiers:interface class
 /// kua
 mixin Layout<T> implements Screen<T> {}
+
+class DefaultLayout<T> extends StatelessWidget with Screen<T>, Layout<T> {
+  final N<T> current;
+
+  final N? tree;
+
+  DefaultLayout({super.key, this.tree, required this.current});
+
+  @override
+  Widget build(BuildContext context) {
+    var left = _NoteTreeView(tree ?? current.root);
+    Pen<T> pen = current.build(context);
+    var center = Scaffold(
+      appBar: AppBar(title: const Text("Flutter Note")),
+      body: ListView(children: [...pen.widgets]),
+    );
+    var right = ListView(
+      shrinkWrap: false,
+      padding: const EdgeInsets.all(20),
+      children: [
+        for (var m in pen.markdownElements)
+          ListTile(
+            title: Row(children: [Text(m.textContent)]),
+          )
+      ],
+    );
+    return Scaffold(
+      body: Row(
+        children: [
+          left,
+          Expanded(child: center),
+          Expanded(child: right),
+        ],
+      ),
+    );
+  }
+
+  @override
+  String get location => current.path;
+
+  @override
+  Rule<T> get rule => current;
+}
+
+class _NoteTreeView extends StatefulWidget {
+  final N root;
+
+  const _NoteTreeView(
+    this.root, {
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  State<_NoteTreeView> createState() => _NoteTreeViewState();
+}
+
+class _NoteTreeViewState extends State<_NoteTreeView> {
+  _NoteTreeViewState();
+
+  @override
+  Widget build(BuildContext context) {
+    // 一页一个链接
+    Widget link(N note) {
+      IconData titleIcon = note.isLeaf ? Icons.remove : Icons.keyboard_arrow_down;
+      click() {
+        // 还未用上这个展开状态，还没想好怎么让ListView模仿树节点的展开和关闭
+        note.extend = !note.extend;
+        NavigatorV2.of(context).push(note.path);
+      }
+
+      return ListTile(
+        trailing: Icon(note.isLeaf ? Icons.open_in_new : null),
+        title: Row(children: [Icon(titleIcon), Text(note.name)]),
+        onTap: note.hasPage ? click : null,
+        // 是否选中
+        selected: false,
+        //---------------下面是外观调整
+        //更紧凑布局
+        dense: false,
+        visualDensity: const VisualDensity(horizontal: -4, vertical: -4),
+        contentPadding: EdgeInsets.only(left: 20 * (note.level - 1).toDouble()),
+      );
+    }
+
+    var notes = widget.root.toList(includeThis: false);
+    return Drawer(
+      child: ListView(
+        shrinkWrap: false,
+        padding: const EdgeInsets.all(20),
+        children: notes.map((e) => link(e)).toList(),
+      ),
+    );
+  }
+}
+
+// 在Note上扩展出UI相关的字段，比如目录树的点开状态`extend`
+extension _TreeViewNote on N {
+  static const _extendAttrName = "note.layout._TreeViewNote.extend";
+
+  //展开状态
+  bool get extend {
+    if (isLeaf) {
+      return false;
+    }
+    Object? result = attributes[_extendAttrName];
+    return result == null ? false : result as bool;
+  }
+
+  set extend(bool extend) {
+    if (isLeaf) {
+      return;
+    }
+    attributes[_extendAttrName] = extend;
+  }
+}
